+++ libs/data-mapper-v2/src/ui/DMReactFlow.tsx
@@ -1,0 +1,233 @@
+import type { AppDispatch, RootState } from '../core/state/Store';
+import { useEffect, useMemo, useRef, useCallback, useState } from 'react';
+import { useSelector, useDispatch } from 'react-redux';
+import type { Connection, Node, Edge, ConnectionLineComponent, NodeDragHandler } from 'reactflow';
+import ReactFlow, { addEdge, useReactFlow } from 'reactflow';
+import { reactFlowStyle, useStaticStyles, useStyles } from './styles';
+import SchemaNode from '../components/common/reactflow/SchemaNode';
+import ConnectionLine from '../components/common/reactflow/ConnectionLine';
+import ConnectedEdge from '../components/common/reactflow/ConnectedEdge';
+import type { ConnectionAction } from '../core/state/DataMapSlice';
+import { makeConnection, updateReactFlowEdges, updateReactFlowNodes } from '../core/state/DataMapSlice';
+import { FunctionNode } from '../components/common/reactflow/FunctionNode';
+import { useDrop } from 'react-dnd';
+
+interface DMReactFlowProps {
+  setIsMapStateDirty?: (isMapStateDirty: boolean) => void;
+  updateCanvasBounds: () => void;
+  canvasBounds: DOMRect | undefined;
+}
+
+export const DMReactFlow = ({ setIsMapStateDirty, updateCanvasBounds, canvasBounds }: DMReactFlowProps) => {
+  useStaticStyles();
+  const styles = useStyles();
+  const reactFlowInstance = useReactFlow();
+  const ref = useRef<HTMLDivElement | null>(null);
+  const dispatch = useDispatch<AppDispatch>();
+  const [allNodes, setAllNodes] = useState<Node[]>([]);
+
+  const resizeObserver = useMemo(() => new ResizeObserver((_) => updateCanvasBounds()), [updateCanvasBounds]);
+
+  const { nodes, edges, functionNodes } = useSelector((state: RootState) => state.dataMap.present.curDataMapOperation);
+
+  useEffect(() => {
+    const newNodes: Node[] = Object.entries(functionNodes).map((node) => ({
+      id: node[0],
+      type: 'functionNode',
+      data: { functionData: node[1] },
+      position: node[1].position || { x: 10, y: 200 },
+      draggable: true,
+    }));
+    setAllNodes(nodes.concat(newNodes));
+  }, [nodes, functionNodes]);
+
+  const isMapStateDirty = useSelector((state: RootState) => state.dataMap.present.isDirty);
+
+  const nodeTypes = useMemo(
+    () => ({
+      schemaNode: SchemaNode,
+      functionNode: FunctionNode,
+    }),
+    []
+  );
+
+  const edgeTypes = useMemo(
+    () => ({
+      connectedEdge: ConnectedEdge,
+    }),
+    []
+  );
+
+  const dispatchEdgesAndNodes = useCallback(
+    (updatedEdges: Edge[], updatedNodes: Node[]) => {
+      const allNodeIds = [...updatedEdges.map((edge) => edge.source), ...updatedEdges.map((edge) => edge.target)];
+
+      const newNodes = [
+        ...updatedNodes.map((node) => ({
+          ...node,
+          data: { ...node.data, isConnected: allNodeIds.indexOf(node.id) > -1 },
+        })),
+      ];
+
+      dispatch(updateReactFlowEdges(updatedEdges));
+
+      dispatch(updateReactFlowNodes(newNodes));
+    },
+    [dispatch]
+  );
+
+  const dispatchMakeConnection = useCallback(
+    (connection: Connection) => {
+      const connectionAction: ConnectionAction = {
+        reactFlowSource: connection.source ?? '',
+        reactFlowDestination: connection.target ?? '',
+      };
+      dispatch(makeConnection(connectionAction));
+    },
+    [dispatch]
+  );
+
+  const onEdgeConnect = useCallback(
+    (connection: Connection) => {
+      const newEdges = addEdge(
+        {
+          ...connection,
+          type: 'connectedEdge',
+          updatable: 'target',
+          focusable: true,
+          deletable: true,
+        },
+        edges
+      );
+
+      dispatchMakeConnection(connection);
+      dispatchEdgesAndNodes(newEdges, nodes);
+    },
+    [edges, nodes, dispatchEdgesAndNodes, dispatchMakeConnection]
+  );
+
+  const onEdgeUpdate = useCallback(
+    (oldEdge: Edge, newConnection: Connection) => {
+      const newEdges = addEdge(
+        {
+          ...newConnection,
+          type: 'connectedEdge',
+          updatable: 'target',
+          focusable: true,
+          deletable: true,
+        },
+        edges.filter((edge) => edge.id !== oldEdge.id)
+      );
+
+      dispatchEdgesAndNodes(newEdges, nodes);
+    },
+    [edges, nodes, dispatchEdgesAndNodes]
+  );
+
+  const isValidConnection = useCallback(
+    (connection: Connection) => {
+      return !edges.find((edge) => edge.source === connection.source && edge.target === connection.target);
+    },
+    [edges]
+  );
+
+  useEffect(() => {
+    if (ref?.current) {
+      resizeObserver.observe(ref.current);
+      updateCanvasBounds();
+    }
+
+    return () => {
+      resizeObserver.disconnect();
+    };
+  }, [ref, resizeObserver, updateCanvasBounds]);
+
+  // NOTE: Putting this useEffect here for vis next to onSave
+  useEffect(() => {
+    if (setIsMapStateDirty) {
+      setIsMapStateDirty(isMapStateDirty);
+    }
+  }, [isMapStateDirty, setIsMapStateDirty]);
+
+  const [, drop] = useDrop(
+    () => ({
+      accept: 'function',
+      drop: (item, monitor) => {
+        const xyPosition = monitor.getClientOffset();
+        if (xyPosition) {
+          if (reactFlowInstance) {
+            const position = reactFlowInstance.screenToFlowPosition({
+              x: xyPosition.x,
+              y: xyPosition.y,
+            });
+            // middle of node is placed where pointer is
+            position.x -= 20;
+            position.y -= 20;
+            return { position };
+          }
+          return { position: xyPosition };
+        }
+        return { position: { x: 0, y: 0 } };
+      },
+    }),
+    [reactFlowInstance]
+  );
+
+  const onFunctionNodeDrag: NodeDragHandler = (_event, node, _nodes) => {
+    const unaffectedNodes = allNodes.filter((nodeFromState) => nodeFromState.id !== node.id);
+    setAllNodes([...unaffectedNodes, node]);
+  };
+
+  //   const onFunctionNodeDragStop: NodeDragHandler = (event, node, _nodes) => {
+  //     const positionMetadata: FunctionPositionMetadata = {
+  //       targetKey: currentTargetSchemaNode?.key || '',
+  //       position: node.position,
+  //     };
+  //     dispatch(updateFunctionPosition({ id: node.id, positionMetadata }));
+  //   };
+
+  return (
+    <div ref={ref} id="editorView" className={styles.canvasWrapper}>
+      <ReactFlow
+        ref={drop}
+        nodes={allNodes}
+        edges={edges}
+        selectNodesOnDrag={false}
+        onlyRenderVisibleElements={false}
+        zoomOnScroll={false}
+        zoomOnPinch={false}
+        nodesConnectable={true}
+        zoomOnDoubleClick={false}
+        nodeTypes={nodeTypes}
+        edgeTypes={edgeTypes}
+        preventScrolling={false}
+        minZoom={1}
+        elementsSelectable={false}
+        maxZoom={1}
+        autoPanOnConnect={false}
+        snapToGrid={true}
+        panOnScroll={false}
+        panOnDrag={false}
+        style={reactFlowStyle}
+        proOptions={{
+          account: 'paid-sponsor',
+          hideAttribution: true,
+        }}
+        onNodeDrag={onFunctionNodeDrag}
+        //onNodeDragStop={onFunctionNodeDragStop}
+        isValidConnection={isValidConnection}
+        onConnect={onEdgeConnect}
+        onEdgeUpdate={onEdgeUpdate}
+        connectionLineComponent={ConnectionLine as ConnectionLineComponent | undefined}
+        translateExtent={
+          canvasBounds
+            ? [
+                [0, 0],
+                [canvasBounds.right, canvasBounds.bottom],
+              ]
+            : undefined
+        }
+      />
+    </div>
+  );
+};
